+++ crates/oxc_napi/Cargo.toml
@@ -33,6 +33,3 @@ napi-derive = { workspace = true }
 
 [build-dependencies]
 napi-build = { workspace = true }
-
-[package.metadata.cargo-shear]
-ignored = ["napi"]
