+++ crates/oxc_napi/src/lib.rs
@@ -1,8 +1,10 @@
 mod comment;
 mod error;
+mod js_reg_exp;
 
 pub use comment::*;
 pub use error::*;
+pub use js_reg_exp::*;
 
 use oxc_ast::{CommentKind, ast::Program};
 use oxc_ast_visit::utf8_to_utf16::Utf8ToUtf16;
